extract common Toolbar component from filter toolbar
